[domain] create repository package & move repository class
