+++ src/main/java/com/humming/repository/CategoryRepository.java
@@ -1,5 +1,6 @@
-package com.humming.core.concert;
+package com.humming.repository;
 
+import com.humming.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
