+++ src/main/java/com/humming/repository/ReservationRepository.java
@@ -1,5 +1,6 @@
-package com.humming.core.reservation;
+package com.humming.repository;
 
+import com.humming.domain.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
